+++ Examples/Example-iOS/Source/AppDelegate.m
@@ -23,51 +23,18 @@
 
 @implementation AppDelegate
 
-@synthesize window = _window;
-
 - (BOOL)application:(UIApplication *)application
     didFinishLaunchingWithOptions:(NSDictionary *)launchOptions {
-  UIWindow *window = [[UIWindow alloc] initWithFrame:[UIScreen mainScreen].bounds];
-  UIViewController *mainViewController =
-      [[GTMAppAuthExampleViewController alloc] init];
-  window.rootViewController = mainViewController;
-
-  _window = window;
-  [_window makeKeyAndVisible];
-
   return YES;
 }
 
-/*! @brief Handles inbound URLs. Checks if the URL matches the redirect URI for a pending
-        AppAuth authorization request.
- */
-- (BOOL)application:(UIApplication *)app
-            openURL:(NSURL *)url
-            options:(NSDictionary<NSString *, id> *)options {
-  // Sends the URL to the current authorization flow (if any) which will process it if it relates to
-  // an authorization response.
-  if ([_currentAuthorizationFlow resumeExternalUserAgentFlowWithURL:url]) {
-    _currentAuthorizationFlow = nil;
-    return YES;
-  }
+#pragma mark - UISceneSession lifecycle
 
-  // Your additional URL handling (if any) goes here.
-
-  return NO;
-}
-
-/*! @brief Forwards inbound URLs for iOS 8.x and below to @c application:openURL:options:.
-    @discussion When you drop support for versions of iOS earlier than 9.0, you can delete this
-        method. NB. this implementation doesn't forward the sourceApplication or annotations. If you
-        need these, then you may want @c application:openURL:options to call this method instead.
- */
-- (BOOL)application:(UIApplication *)application
-              openURL:(NSURL *)url
-    sourceApplication:(NSString *)sourceApplication
-           annotation:(id)annotation {
-  return [self application:application
-                   openURL:url
-                   options:@{}];
+- (UISceneConfiguration *)application:(UIApplication *)application
+    configurationForConnectingSceneSession:(UISceneSession *)connectingSceneSession
+                                   options:(UISceneConnectionOptions *)options {
+  return [[UISceneConfiguration alloc] initWithName:@"Default Configuration"
+                                         sessionRole:connectingSceneSession.role];
 }
 
 @end