+++ Examples/Example-iOS/Source/SceneDelegate.m
@@ -1,0 +1,48 @@
+/*! @file SceneDelegate.m
+    @brief GTMAppAuth SDK iOS Example
+    @copyright
+        Copyright 2026 Google LLC
+    @copydetails
+        Licensed under the Apache License, Version 2.0 (the "License");
+        you may not use this file except in compliance with the License.
+        You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+        Unless required by applicable law or agreed to in writing, software
+        distributed under the License is distributed on an "AS IS" BASIS,
+        WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+        See the License for the specific language governing permissions and
+        limitations under the License.
+ */
+#import "SceneDelegate.h"
+
+@import AppAuth;
+
+#import "GTMAppAuthExampleViewController.h"
+
+@implementation SceneDelegate
+
+- (void)scene:(UIScene *)scene
+    willConnectToSession:(UISceneSession *)session
+                 options:(UISceneConnectionOptions *)connectionOptions {
+  if (![scene isKindOfClass:[UIWindowScene class]]) {
+    return;
+  }
+  UIWindowScene *windowScene = (UIWindowScene *)scene;
+  self.window = [[UIWindow alloc] initWithWindowScene:windowScene];
+  self.window.rootViewController = [[GTMAppAuthExampleViewController alloc] init];
+  [self.window makeKeyAndVisible];
+}
+
+- (void)scene:(UIScene *)scene openURLContexts:(NSSet<UIOpenURLContext *> *)URLContexts {
+  UIOpenURLContext *context = URLContexts.anyObject;
+  if (!context) {
+    return;
+  }
+  if ([self.currentAuthorizationFlow resumeExternalUserAgentFlowWithURL:context.URL]) {
+    self.currentAuthorizationFlow = nil;
+  }
+}
+
+@end